Pitch bend
Been experimenting with a form of pitch bend. 
Essentially when you play a note nothing will happen for a short period of time, then pitch bending will turn on and moving your finger slightly on the note can create a vibrato effect. 
An indicator LED has been added to show the status of the pitch bending on pin 10. The pitch bending can now be enabled/disabled using the right button on the MIDI shield (Pin 2).

Currently there are some issues when transitioning between notes and I don't think there is any real way around that just because of the way MIDI messages are sent.